feat(Dashboard): consolidated outgoing contributions
